New module for comparing with Olly. In this module, the multi-comparison with many files is allowed.

+++ src/v2/org/analysis/apihandle/winapi/kernel32/Kernel32DLLwithoutOption.java
@@ -19,5 +19,5 @@ public interface Kernel32DLLwithoutOption extends StdCallLibrary {
 
 	Kernel32DLLwithoutOption INSTANCE = (Kernel32DLLwithoutOption) Native.loadLibrary("kernel32", Kernel32DLLwithoutOption.class);
 	Kernel32DLLwithoutOption SYNC_INSTANCE = (Kernel32DLLwithoutOption) Native.synchronizedLibrary(INSTANCE);
-	long GetProcAddress(HMODULE hM, String lpProcName);
+	int GetProcAddress(HMODULE hM, String lpProcName);
 }
